Refactor : ajaxUtil.js 신규 작성 및 적용
