+++ src/js/util/ajaxUtil.js
@@ -1,0 +1,10 @@
+export async function get(url,errMsg){
+    const response = await fetch(url);
+    let body = [];
+    if (response.ok) { 
+      body = await response.json();
+    } else {
+      alert(errMsg + " : " + response.status);
+    }
+    return body;
+}
